refactor(swiftui): split DocumentInspector into focused files (#1703)
Extract four standalone types from the 1807-line monolith into sibling
files, matching the OntologyBrowser split pattern (commit 8ca6c431):

- InspectorTab.swift (64 lines) — InspectorTab enum + Notification.Name ext
- DisplayAttributesStrip.swift (502 lines) — attribute strip V2
- ArtifactPanel.swift (535 lines) — CatalogueArtifactPreviews + ArtifactPanel
- DocumentInspectorContentV2.swift (323 lines) — content V2 + persistPageContent

DocumentInspector.swift drops from 1807 → 381 lines; removes the top-level
// swiftlint:disable file_length comment (no longer needed). All new files
registered via scripts/add-swift-file.rb. Zero logic or visual changes.
